+++ Source/ISHDeploy/Business/Operations/ISHUIEventMonitorTab/SetISHUIEventMonitorTabOperation.cs
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (c) 2014 All Rights Reserved by the SDL Group.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -24,7 +24,7 @@ namespace ISHDeploy.Business.Operations.ISHUIEventMonitorTab
 	/// Sets Event Monitor Tab.
 	/// </summary>
 	/// <seealso cref="ISHDeploy.Business.Operations.IOperation" />
-	public class SetISHUIEventMonitorTabOperation : BasePathsOperation, IOperation
+	public class SetISHUIEventMonitorTabOperation : BaseOperationPaths, IOperation
 	{
 		/// <summary>
 		/// The actions invoker
@@ -42,7 +42,7 @@ namespace ISHDeploy.Business.Operations.ISHUIEventMonitorTab
 		{
 			_invoker = new ActionInvoker(logger, "Setting of Event Monitor Tab");
 
-			_invoker.AddAction(new SetNodeAction(logger, EventMonitorMenuBarXml.Path, string.Format(EventMonitorMenuBarXml.EventMonitorTab, menuItem.Label), menuItem));
+			_invoker.AddAction(new SetNodeAction(logger, EventMonitorMenuBarXmlPath, string.Format(EventMonitorMenuBarXml.EventMonitorTab, menuItem.Label), menuItem));
 		}
 
 		/// <summary>